refactor: put all cli commands to same place
