+++ cli/__init__.py
@@ -1,0 +1,2 @@
+import cli.db
+import cli.qa
