+++ server/__init__.py
@@ -59,17 +59,22 @@ app.jinja_env.filters.update(
     max=max,
 )
 
+# These must be done after `app`` is created as they depend on `app`
 
+# prepares all auth clients
 import server.services.auth  # noqa
-import server.models  # noqa
 
+# registers controller converters
 from server.controllers import ExamConverter, OfferingConverter  # noqa
 app.url_map.converters['exam'] = ExamConverter
 app.url_map.converters['offering'] = OfferingConverter
 
+# registers base controllers (depends on converters, so must be done after)
 import server.views  # noqa
 
+# registers blueprint controllers
 from server.controllers import auth_module  # noqa
 app.register_blueprint(auth_module)
 
-import tests.conftest  # noqa
+# registers flask cli commands
+import cli  # noqa
