+++ tests/conftest.py
@@ -50,45 +50,3 @@ def driver():
     driver = webdriver.Chrome(options=options)
     yield driver
     driver.close()
-
-
-@flask_app.cli.command('e2e')
-def run_e2e():
-    print('Running end-to-end tests...')
-    pytest.main(['-s', 'tests/e2e'])
-
-
-@flask_app.cli.command('unit')
-def run_unit():
-    print('Running unit tests...')
-    pytest.main(['-s', 'tests/unit'])
-
-
-@flask_app.cli.command('test')
-def run_all():
-    print('Running all tests...')
-    pytest.main(['-s', 'tests'])
-
-
-@flask_app.cli.command('cov')
-def run_cov():
-    print('Running all tests with coverage...')
-    pytest.main(['-s', '--cov=server', 'tests'])
-
-
-@flask_app.cli.command('audit')
-def run_audit():
-    print('Checking for broken dependencies...')
-    os.system('pip check')
-    print('Checking for outdated dependencies...')
-    os.system('pip list --outdated')
-    print('Auditing with pip-audit...')
-    os.system('pip-audit')
-    print('Auditing with safety...')
-    os.system('safety check')
-
-
-@flask_app.cli.command('lint')
-def run_lint():
-    print('Running flake8 linter...')
-    os.system('flake8 server tests')
